+++ spec/rspec/core/command_line_spec.rb
@@ -95,6 +95,8 @@ module RSpec::Core
         # this is necessary to ensure that color works correctly on windows
         config.should_receive(:error_stream=).ordered
         config.should_receive(:output_stream=).ordered
+        config.should_receive(:force).with(:default_path => anything).ordered
+        config.should_receive(:force).with(:order => anything).ordered
         config.should_receive(:force).with(:exclusion_filter => anything).ordered
         config.should_receive(:force).with(:color => true).ordered
         command_line.run(err, out) rescue nil